split out snapstate/handlers.go
